Replace L.browser.touch with this._map.tap
The L.browser.touch variable seems to be unreliable to distinguish
touch- from non-touch devices. At least on Microsoft's
Internet Explorer 11 and Edge this causes an error because
L.browser.touch is truthy but this._map.tap does not exist
and this._map.tap.disable() therefore raises an error.

this._map.tap seems to be the better method for detecting touch
capabilities here.